requirements.txt
# requirements.txt
flask
python-dotenv
PyPDF2
openai
faiss-cpu
